runtime_probe: Update usage of ec::I2cReadCommand
Update ec::I2cReadCommand::Data usage to match the function signature
change.

BUG=b:298160433
TEST=FEATURES=test emerge-brya runtime_probe

Cq-Depend: chromium:4843733
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4853357
Commit-Queue: Kevin Lin <kevinptt@chromium.org>
Tested-by: Kevin Lin <kevinptt@chromium.org>

+++ runtime_probe/functions/ec_i2c_test.cc
@@ -44,7 +44,7 @@ class EcI2cFunctionTest : public BaseFunctionTest {
     }
 
     MOCK_METHOD(bool, Run, (int), (override));
-    MOCK_METHOD(uint16_t, Data, (), (const override));
+    MOCK_METHOD(uint32_t, Data, (), (const override));
     MOCK_METHOD(uint32_t, Result, (), (const override));
     MOCK_METHOD(uint8_t, I2cStatus, (), (const override));
   };
